@@ -5,8 +5,14 @@
   "description": "Magento Cron Scheduler",
   "homepage": "https://github.com/AOEpeople/Aoe_Scheduler",
   "require": {
-    "php": ">=5.3",
+    "php": "^8.0",
+    "ext-mbstring": "*",
+    "ext-pdo": "*",
+    "ext-posix": "*",
     "magento-hackathon/magento-composer-installer": "*"
+  },
+  "require-dev": {
+    "phpunit/phpunit": "^9.6"
   },
   "authors": [
     {
