@@ -5,7 +5,7 @@ class CronGroupsWhiteListAlwaysTest extends AbstractTest
 
     protected $groups = array();
 
-    protected function setUp()
+    protected function setUp(): void
     {
         parent::setUp();
 
