@@ -2,10 +2,9 @@
 
 class CronGroupsWhiteListTest extends AbstractTest
 {
-
     protected $groups = array();
 
-    protected function setUp()
+    protected function setUp(): void
     {
         parent::setUp();
 
